python_backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
dnspython>=2.4.2
python-whois>=0.8.0
geoip2>=4.8.0
maxminddb>=2.2.0
cryptography>=42.0.0
bcrypt>=4.1.2
pyjwt>=2.8.0
